Dynamic capture header
The amount of rows before the profile differs between metaphlan3 and 4. Therefore taxpasta was not working on metaphlan4 output under the current circumstances. Since the header is formatted the same between both versions, a quick scan for the first field of the header could ensure the table is read from the right line in both formats.